refactor: 스터디 출석 시 포인트 지급하는 로직을 이벤트 리스너에서 일반 메서드로 이동

+++ src/test/java/aegis/server/domain/study/service/StudyPointRewardTest.java
@@ -112,7 +112,7 @@ class StudyPointRewardTest extends IntegrationTestWithoutTransactional {
 
         PointAccount refreshed =
                 pointAccountRepository.findByMemberId(instructor.getId()).orElseThrow();
-        assertEquals(30, refreshed.getBalance().intValue());
+        assertEquals(BigDecimal.valueOf(30), refreshed.getBalance());
     }
 
     @Test
